split up the communication logic from the esp interface
